Revert "compositor?"
This reverts commit d11801d9dcb6b01a25dce92d4ed2fd427b776947.

+++ intern/mikktspace/mikk_float3.hh
@@ -64,7 +64,7 @@ struct float3 {
 
   float length() const
   {
-    return sqrtf(length_squared());
+    return sqrt(length_squared());
   }
 
   static float distance(const float3 &a, const float3 &b)
